requirements.txt
flask
flask-mqtt
llama-mqtt
ConfigParser
pypiwin32
wmi